Move partner commission creation into create-partner-commission workflow.
