+++ apps/web/lib/actions/partners/create-manual-commission.ts
@@ -380,17 +380,15 @@ export const createManualCommissionAction = authActionClient
     const tbRes = await Promise.allSettled(tbEventsToRecord.map((fn) => fn()));
     console.log("Recorded events in Tinybird: ", prettyPrint(tbRes));
 
-    let createdCommissions = 0;
+    let queuedCommissions = 0;
     // create partner commissions (use a for loop to make sure the commissions are created in the correct order)
     // TODO: migrate to use workflow to support bulk creation
     for (const c of commissionsToCreate) {
-      const { commission } = await createPartnerCommission(c);
-      if (commission) {
-        createdCommissions++;
-      }
+      await createPartnerCommission(c);
+      queuedCommissions++;
     }
     console.log(
-      `Created ${createdCommissions} commissions for partner ${partner.email} (${partner.id}) and customer ${customer.email} (${customer.id})`,
+      `Queued ${queuedCommissions} commissions for partner ${partner.email} (${partner.id}) and customer ${customer.email} (${customer.id})`,
     );
 
     waitUntil(
