+++ apps/web/lib/api/conversions/track-lead.ts
@@ -1,6 +1,5 @@
 import { createId } from "@/lib/api/create-id";
 import { DubApiError } from "@/lib/api/errors";
-import { detectAndRecordFraudEvent } from "@/lib/api/fraud/detect-record-fraud-event";
 import { includeTags } from "@/lib/api/links/include-tags";
 import { generateRandomName } from "@/lib/names";
 import { createPartnerCommission } from "@/lib/partners/create-partner-commission";
@@ -17,7 +16,7 @@ import {
 } from "@/lib/zod/schemas/leads";
 import { prisma } from "@dub/prisma";
 import { Link } from "@dub/prisma/client";
-import { nanoid, pick, R2_URL } from "@dub/utils";
+import { nanoid, R2_URL } from "@dub/utils";
 import { waitUntil } from "@vercel/functions";
 import * as z from "zod/v4";
 import { syncPartnerLinksStats } from "../partners/sync-partner-links-stats";
@@ -287,12 +286,12 @@ export const trackLead = async ({
           ]);
           link = updatedLink; // update the link variable to the latest version
 
-          let createdCommission:
-            | Awaited<ReturnType<typeof createPartnerCommission>>
-            | undefined = undefined;
+          let result: Awaited<
+            ReturnType<typeof createPartnerCommission>
+          > | null = null;
 
           if (link.programId && link.partnerId && customer) {
-            createdCommission = await createPartnerCommission({
+            result = await createPartnerCommission({
               event: "lead",
               programId: link.programId,
               partnerId: link.partnerId,
@@ -307,9 +306,6 @@ export const trackLead = async ({
                 },
               },
             });
-
-            const { commission, webhookPartner, programEnrollment } =
-              createdCommission;
 
             await Promise.allSettled([
               executeWorkflows({
@@ -334,17 +330,17 @@ export const trackLead = async ({
               }),
 
               // only run fraud checks if the commission was created
-              commission &&
-                webhookPartner &&
-                detectAndRecordFraudEvent({
-                  program: { id: link.programId },
-                  partner: pick(webhookPartner, ["id", "email", "name"]),
-                  programEnrollment: pick(programEnrollment, ["status"]),
-                  customer: pick(customer, ["id", "email", "name"]),
-                  link: pick(link, ["id"]),
-                  click: pick(clickData, ["url", "referer"]),
-                  event: { id: leadEventId },
-                }),
+              // commission &&
+              //   webhookPartner &&
+              //   detectAndRecordFraudEvent({
+              //     program: { id: link.programId },
+              //     partner: pick(webhookPartner, ["id", "email", "name"]),
+              //     programEnrollment: pick(programEnrollment, ["status"]),
+              //     customer: pick(customer, ["id", "email", "name"]),
+              //     link: pick(link, ["id"]),
+              //     click: pick(clickData, ["url", "referer"]),
+              //     event: { id: leadEventId },
+              //   }),
             ]);
           }
 
@@ -356,7 +352,7 @@ export const trackLead = async ({
                 eventName,
                 link,
                 customer,
-                partner: createdCommission?.webhookPartner,
+                partner: result?.webhookPartner,
                 metadata,
               }),
               workspace,
