+++ apps/web/lib/integrations/shopify/create-sale.ts
@@ -1,5 +1,4 @@
 import { isFirstConversion } from "@/lib/analytics/is-first-conversion";
-import { detectAndRecordFraudEvent } from "@/lib/api/fraud/detect-record-fraud-event";
 import { includeTags } from "@/lib/api/links/include-tags";
 import { syncPartnerLinksStats } from "@/lib/api/partners/sync-partner-links-stats";
 import { executeWorkflows } from "@/lib/api/workflows/execute-workflows";
@@ -11,7 +10,7 @@ import { redis } from "@/lib/upstash";
 import { sendWorkspaceWebhook } from "@/lib/webhook/publish";
 import { transformSaleEventData } from "@/lib/webhook/transform";
 import { prisma } from "@dub/prisma";
-import { nanoid, pick } from "@dub/utils";
+import { nanoid } from "@dub/utils";
 import { waitUntil } from "@vercel/functions";
 import { orderSchema } from "./schema";
 
@@ -132,12 +131,11 @@ export async function createShopifySale({
   ]);
 
   // for program links
-  let createdCommission:
-    | Awaited<ReturnType<typeof createPartnerCommission>>
-    | undefined = undefined;
+  let result: Awaited<ReturnType<typeof createPartnerCommission>> | undefined =
+    undefined;
 
   if (link.programId && link.partnerId) {
-    createdCommission = await createPartnerCommission({
+    result = await createPartnerCommission({
       event: "sale",
       programId: link.programId,
       partnerId: link.partnerId,
@@ -158,8 +156,6 @@ export async function createShopifySale({
         },
       },
     });
-
-    const { webhookPartner, programEnrollment } = createdCommission;
 
     waitUntil(
       Promise.allSettled([
@@ -185,19 +181,19 @@ export async function createShopifySale({
           eventType: "sale",
         }),
 
-        webhookPartner &&
-          detectAndRecordFraudEvent({
-            program: { id: link.programId },
-            partner: pick(webhookPartner, ["id", "email", "name"]),
-            programEnrollment: pick(programEnrollment, ["status"]),
-            customer: {
-              ...pick(customer, ["id", "email", "name"]),
-              isFirstConversion: firstConversionFlag,
-            },
-            link: pick(link, ["id"]),
-            click: pick(saleData, ["url", "referer"]),
-            event: { id: saleData.event_id },
-          }),
+        // webhookPartner &&
+        //   detectAndRecordFraudEvent({
+        //     program: { id: link.programId },
+        //     partner: pick(webhookPartner, ["id", "email", "name"]),
+        //     programEnrollment: pick(programEnrollment, ["status"]),
+        //     customer: {
+        //       ...pick(customer, ["id", "email", "name"]),
+        //       isFirstConversion: firstConversionFlag,
+        //     },
+        //     link: pick(link, ["id"]),
+        //     click: pick(saleData, ["url", "referer"]),
+        //     event: { id: saleData.event_id },
+        //   }),
       ]),
     );
   }
@@ -212,7 +208,7 @@ export async function createShopifySale({
           link,
           clickedAt: customer.clickedAt || customer.createdAt,
           customer,
-          partner: createdCommission?.webhookPartner,
+          partner: result?.webhookPartner,
           metadata: null,
         }),
       }),
